fix the wrong result produced on pitch yaw roll data where it is ok on raspberry pi. added test code and test results. Here is the description: 1. Eigen library (Matrix math C++ generic library) 2. <math.h> library (trigonometry math c++ generic library) 3. operation on double type. The above three features have been verified from TI board against the test result generated from Raspberry Pi. (Take me a whole week to verify every detail) And it turns out the reason preventing the code producing right value on TI board is that an variation of the compiler. The gcc compiler i used in Raspberry Pi but TI compiler is used in TI chip. Some assumptions made in gcc compiler might not pass down to TI compiler instead. Let me explain, there is a class called QuaterClass, and a method return this Class type as well, then we have: quaternion = EKF,update(x,y,z,dt); The LHS is QuaterClass type and RHS returns a QuaterClass type as well. The gcc compiler will automatically generate an overload on operator= to copy the contents of RHS to LHS but TI compiler did not. Nevertheless, it is necessary for programmer to define what operator= does when the compiler handles it with this class. (E.g. in Java, operator= for classes means to pass reference, but in c++ with gcc compiler, operator= means copy to contents over, in c++ with TI compiler, I guess it just does non-sense)

+++ CCS_projects/Overseer/src/TopLevel/StandaloneCopter/OptimusPiCopter/AHRS/Quaternion.cpp
@@ -17,7 +17,7 @@ QuaternionClass::~QuaternionClass() {
 }
 
  void QuaternionClass::normalise(){
-    double norm = sqrt(w*w + x*x + y*y + z*z);
+    float norm = sqrt(w*w + x*x + y*y + z*z);
     w /= norm;
     x /= norm;
     y /= norm;
